# specs/001-comidapp-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ComidApp API
  version: 1.0.0
  description: >
    API REST para ComidApp — plataforma web de pedidos y entrega de comida rápida.
    Todos los endpoints protegidos requieren token JWT en el header Authorization.
  contact:
    name: ComidApp Team
servers:
  - url: http://localhost:3001/api/v1
    description: Desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EstadoPedido:
      type: string
      enum: [RECIBIDO, EN_PREPARACION, EN_CAMINO, ENTREGADO, CANCELADO]
    MetodoPago:
      type: string
      enum: [TARJETA, EFECTIVO]
    EstadoPago:
      type: string
      enum: [PENDIENTE, APROBADO, RECHAZADO, REVERTIDO]
    RolUsuario:
      type: string
      enum: [CLIENTE, ADMIN, REPARTIDOR]

    # ── Shared ───────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [error, mensaje]
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        mensaje:
          type: string
          example: El campo email es requerido
    Paginacion:
      type: object
      required: [pagina, porPagina, total, totalPaginas]
      properties:
        pagina:
          type: integer
          minimum: 1
        porPagina:
          type: integer
          example: 10
        total:
          type: integer
        totalPaginas:
          type: integer

    # ── Usuarios ─────────────────────────────────────────────────────────────
    UsuarioPublico:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
          format: email
        telefono:
          type: string
          nullable: true
        direccionEntrega:
          type: string
          nullable: true
        rol:
          $ref: '#/components/schemas/RolUsuario'

    # ── Locales ──────────────────────────────────────────────────────────────
    Local:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
          example: El Monstruo
        direccion:
          type: string
        barrio:
          type: string
        tiempoEstimadoMin:
          type: integer
          example: 30
        abierto:
          type: boolean
          description: Indica si el local está dentro de su horario de atención ahora
...
    ItemCarrito:
      type: object
      properties:
        id:
          type: integer
        itemMenuId:
          type: integer
        nombre:
          type: string
        precio:
          type: number
          format: double
        disponible:
          type: boolean
          description: false si el producto fue desactivado mientras estaba en el carrito
        cantidad:
          type: integer
          minimum: 1
        subtotal:
          type: number
          format: double
    ResumenCarrito:
      type: object
      properties:
        localId:
          type: integer
        localNombre:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemCarrito'
        subtotal:
          type: number
          format: double
        costoEnvio:
          type: number
          format: double
        total:
          type: number
          format: double
        tieneItemsNoDisponibles:
          type: boolean
          description: true si algún item fue desactivado; bloquea la confirmación

    # ── Pedidos ──────────────────────────────────────────────────────────────
    ItemPedido:
      type: object
      properties:
        productoId:
          type: integer
        nombreProducto:
          type: string
        precioUnitario:
          type: number
          format: double
        cantidad:
          type: integer
        subtotal:
          type: number
          format: double
    Pedido:
      type: object
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        localId:
          type: integer
        localNombre:
          type: string
        repartidorId:
          type: integer
          nullable: true
        estado:
          $ref: '#/components/schemas/EstadoPedido'
        metodoPago:
          $ref: '#/components/schemas/MetodoPago'
        precioTotal:
          type: number
          format: double
        costoEnvio:
          type: number
          format: double
        direccionEntrega:
          type: string
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        pago:
          $ref: '#/components/schemas/Pago'

    # ── Pagos ────────────────────────────────────────────────────────────────
    Pago:
      type: object
      properties:
        id:
          type: integer
        estado:
          $ref: '#/components/schemas/EstadoPago'
        metodo:
          $ref: '#/components/schemas/MetodoPago'
        monto:
          type: number
          format: double
        externalPaymentId:
          type: string
          nullable: true
paths:
...
  /carrito/items:
    post:
      tags: [Carrito]
      summary: Agregar un ítem al carrito (FR-007, FR-007b)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemMenuId, cantidad]
              properties:
                itemMenuId:
                  type: integer
                cantidad:
                  type: integer
                  minimum: 1
                limpiarCarritoSiOtroLocal:
                  type: boolean
                  default: false
                  description: >
                    Si true y el item pertenece a otro local, limpia el carrito actual
                    y comienza uno nuevo. Si false y hay conflicto de local, devuelve 409.
      responses:
        '200':
          description: Carrito actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumenCarrito'
        '409':
          description: El carrito ya contiene items de otro local y limpiarCarritoSiOtroLocal=false
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: LOCAL_CONFLICT
                  localActual:
                    $ref: '#/components/schemas/Local'
                  localNuevo:
                    $ref: '#/components/schemas/Local'
  /carrito/items/{itemMenuId}:
    put:
      tags: [Carrito]
      summary: Modificar cantidad de un ítem en el carrito (FR-008)
      security:
        - bearerAuth: []
      parameters:
        - name: itemMenuId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cantidad]
              properties:
                cantidad:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Carrito actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumenCarrito'
        '404':
          description: Item no encontrado en el carrito

    delete:
      tags: [Carrito]
      summary: Eliminar un ítem del carrito (FR-009)
      security:
        - bearerAuth: []
      parameters:
        - name: itemMenuId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Carrito actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumenCarrito'
  /carrito:
    delete:
      tags: [Carrito]
      summary: Vaciar el carrito completo
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Carrito vaciado

  # ══════════════════════════════════════════════════════════
  # PEDIDOS
  # ══════════════════════════════════════════════════════════
  /pedidos:
    post:
      tags: [Pedidos]
      summary: Confirmar pedido desde el carrito activo (FR-011, FR-012, FR-013, FR-033)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [metodoPago]
              properties:
                metodoPago:
                  $ref: '#/components/schemas/MetodoPago'
                direccionEntrega:
                  type: string
                  description: Opcional; si no se envía, usa la del perfil del usuario
      responses:
        '201':
          description: Pedido creado con estado RECIBIDO
          content:
            application/json:
              schema:
                type: object
                properties:
                  pedido:
                    $ref: '#/components/schemas/Pedido'
                  checkoutUrl:
                    type: string
                    nullable: true
                    description: URL de Mercado Pago si metodoPago=TARJETA; null si EFECTIVO
        '400':
          description: Carrito vacío, productos no disponibles o local cerrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [CARRITO_VACIO, PRODUCTOS_NO_DISPONIBLES, LOCAL_CERRADO]
                  detalle:
                    type: string
                  itemsNoDisponibles:
                    type: array
                    items:
                      type: integer
                    description: IDs de item_menu no disponibles (solo para PRODUCTOS_NO_DISPONIBLES)
        '409':
          description: No hay repartidores disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/usuarios:
    get:
      tags: [Admin]
      summary: Listar todos los usuarios (FR-027)
      security:
        - bearerAuth: []
      parameters:
        - name: rol
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RolUsuario'
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Lista de usuarios paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuarios:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsuarioPublico'
                  paginacion:
                    $ref: '#/components/schemas/Paginacion'
...
  /admin/usuarios/{usuarioId}:
    delete:
      tags: [Admin]
      summary: Eliminar usuario del sistema (FR-028)
      security:
        - bearerAuth: []
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usuario eliminado
        '403':
          description: No se puede eliminar al propio admin

  # ══════════════════════════════════════════════════════════
  # REPARTIDOR
  # ══════════════════════════════════════════════════════════
...